[WOR-799] Add ability to attach to existing azure resources (#50)
Context

We need a way to create a landing zone for a limited set of critical path E2E tests (think: create workspace, delete workspace, etc.). To do so, we'll need a landing zone, but do not want to wait the ~10 minutes to spin up a "real" landing zone. In a testing configuration, the ability to "attach" to pre-deployed landing zone resources should suffice.

Changes

Adds a parameter to the creation job, attach. When present (and the LZ service configuration allows it), this will skip the creation of landing zone resources in the MRG. The flight still goes through the steps of checking the billing profile permissions and creating a Sam resource.
Adds a configuration LandingZoneTestingConfiguration. This controls whether the LZS allows the attachment functionality. I expect this will be off everywhere except for BEEs in a testing context.