[WOR-861] Proceed with db and sam record deletion if the backing resource group is inaccessible (#55)
Context

If the MRG backing a landing zone is removed, deleting the landing zone results in an exception when the deletion code goes to pull in the resources for the landing zone.

This PR

Handles the AuthorizationFailed error from Azure and allows deletion to proceed. Azure gives this error back when the MRG is deleted rather than a 404-type error. Given that the IAM on the MRG is not alterable by users due to deny assignments, we can assume that the MRG has been removed and can proceed forward with deletion.