[WOR-952] Exclude vulnerable version of json-smart (#81)
json-smart 2.4.8 is still pinging the sourceclear scans for this library even after upgrading to the latest azure-identity. This is likely because an upstream dep of azure-identity is still pulling in 2.4.8 transitively even though azure-identity has upgraded itself. I'm explicitly excluding the dep in both service and testharness and pulling in 2.4.10. It looks like the next azure-sdk should fix this issue so we should look into removing this workaround once that is released in May.