[WOR-1012] Wire up long term storage data export from protected LZ LA workspace (#112)
Why

We need to preserve logs from protected data landing zones for up to three years in a central, Broad-managed subscription. To do so, we are creating a data export from the Log Analytics workspace in the LZ to one of the storage account's that was provisioned in this Terraform.

This PR

Adds a configuration entry for the administrative storage accounts. In a previous rev of this PR, I had taken the approach of fetching the storage accounts from a config'ed subscription and finding the correct one's region. This proved to be quite cumbersome -- I prefer the simpler config approach for now.
Adds a step to create the actual data export wire up to the log analytics workspace.