[WOR-1081] Add retries when creating alert rules (#159)
Sentinel alert rules intermittently fail to create with errors like this:
{\\\"code\\\":\\\"SemanticError\\\",\\\"message\\\":\\\"A semantic error occurred.\\\",\\\"innererror\\\":{\\\"code\\\":\\\"SEM0104\\\",\\\"message\\\":\\\"Operator source expression should be table or column\\\"}}}};\"}}

Googling turns up very little and this is not consistently reproducible. I created a test locally that created a minimal landing zone, attached a sentinel then attempted to create the alert rules. This did repro the issue more often but still not 100% of the time.

My guess is that we are creating the alert rules too soon after creating the sentinel instance. When I introduce retry logic into the flight, the alert rules create successfully (when run via the ITs in CI; running the above adhoc test does not exercise stairway retry logic but instead runs each step manually).

This PR

Adds the ability for BaseResourceCreateStep subclasses to override a maybeHandleManagementException method, which I've done for the CreateSentinelAlertRulesStep. The override in this case handles 400 BadRequest errors, which we do not generally want to do but is the error condition for when these alert rules fail out.