[WOR-1299] Use federated credentials when connecting to Azure in CI (#284)
The preferred mechanism for connecting to Azure in GitHub is using a federated credential. This should mitigate the leakage of our client secret in PRs, logs, etc.

This PR

Modifies the build-and-test GitHub CI action to login to Azure using a previously setup federated credential.
Switches the Azure connection code in the service to use the DefaultAzureCredentialBuilder and removes the client secret configuration entries. The credentials must now be supplied via environment variables or available via the Azure CLI as documented here and here.