Upgrade TCL 1.1.0, streamline OTEL deps
This TCL upgrade includes an update to a vulnerable transitive dependency.
We discovered when working in WSM that the vast majority of our defined OTEL deps are unnecessary.
But, in getting rid of them I found that I needed to explicitly add in the guava dependency for TCL to work, we previously got it transitively through an otherwise unnecessary OTEL dependency.