[PF-3000] TCL 1.1.4, k8s client 20.0.1-legacy
Move LZS back to using the legacy version of the k8s client to match what is in TCL 1.1.4.  The non-legacy version yielded a breaking change in TCL's KubePodListener.

Instruct Dependabot to ignore k8s client upgrades.

Finished wiring in srcclr plug-in (an additional block of config was required to run it), added instructions for running scans off of local code and feature branches.