[WOR-1448] Upgrade TCL 1.1.1 -> 1.1.6
This TCL version includes KubePodListener fixes to work with the non-legacy k8s client.  Related PR: DataBiosphere/terra-common-lib#174

Instruct Dependabot to ignore k8s client upgrades.  The client is publishing legacy and non-legacy versions, and Dependabot tries to update to the latest published version.

Finished wiring in srcclr plug-in (an additional block of config was required to run it), added instructions for running scans off of local code and feature branches.