Use credentials in env vars to auth w/Azure during ITs (#443)
The OIDC token minted by GitHub is good for ~10 minutes. The downstream jobs in the Landing Zone Service ITs fail out when they go to get a fresh token after that expiration. This seems to have regressed recently and we cannot control the expiration on the GH token.

This PR switches the LZ CI to use well-known Azure env vars (tenant + client_id + secret) instead of a federated cred obtained via the Azure CLI.